feat: search

+++ src/main/java/ru/yandex/practicum/filmorate/storage/db/UserDbStorage.java
@@ -10,18 +10,31 @@ import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.interfaces.UserStorage;
 
 import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
+import java.util.*;
 
 @Repository
 public class UserDbStorage implements UserStorage {
+    private static final String GET_USER_ID = "SELECT user_id FROM users WHERE user_id=?";
     private final JdbcTemplate jdbcTemplate;
 
     public UserDbStorage(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
+    }
+
+    private static User userMap(SqlRowSet srs) {
+        int id = srs.getInt("user_id");
+        String name = srs.getString("user_name");
+        String login = srs.getString("login");
+        String email = srs.getString("email");
+        LocalDate birthday = Objects.requireNonNull(srs.getTimestamp("birthday"))
+                .toLocalDateTime().toLocalDate();
+        return User.builder()
+                .id(id)
+                .name(name)
+                .login(login)
+                .email(email)
+                .birthday(birthday)
+                .build();
     }
 
     public Collection<User> getUsers() {
@@ -32,6 +45,11 @@ public class UserDbStorage implements UserStorage {
             users.add(userMap(srs));
         }
         return users;
+    }
+
+    @Override
+    public boolean containsUser(int id) {
+        return jdbcTemplate.queryForRowSet(GET_USER_ID, id).next();
     }
 
     @Override
@@ -65,10 +83,12 @@ public class UserDbStorage implements UserStorage {
     }
 
     @Override
-    public String delete(int userId) {
-        String sqlQuery = "DELETE FROM users WHERE user_id = ?";
-        jdbcTemplate.update(sqlQuery, userId);
-        return sqlQuery;
+    public void delete(int userId) {
+        String sqlQuery = "DELETE FROM users WHERE user_id = " + userId;
+        int numberModifiedRows = jdbcTemplate.update(sqlQuery);
+        if (numberModifiedRows < 1) {
+            throw new ObjectNotFoundException(LogMessagesUsers.USER_NO_FOUND_WITH_ID.getMessage());
+        }
     }
 
     public User getById(Integer userId) {
@@ -88,12 +108,13 @@ public class UserDbStorage implements UserStorage {
     }
 
     public void removeFriend(int userId, int friendId) {
-        String sqlQuery = "DELETE friends "
+        String sqlQuery = "DELETE FROM friends "
                 + "WHERE user_id = ? AND friend_id = ?";
         jdbcTemplate.update(sqlQuery, userId, friendId);
     }
 
     public List<User> getFriends(int userId) {
+        getById(userId);
         List<User> friends = new ArrayList<>();
         String sqlQuery = "SELECT * FROM users "
                 + "WHERE users.user_id IN (SELECT friend_id from friends "
@@ -108,7 +129,7 @@ public class UserDbStorage implements UserStorage {
     public List<User> getCommonFriends(int friend1, int friend2) {
         List<User> commonFriends = new ArrayList<>();
         String sqlQuery = "SELECT * FROM users "
-                + "WHERE users.user_id IN (SELECT friend_id from friends "
+                + "WHERE users.user_id IN (SELECT friend_id FROM friends "
                 + "WHERE user_id IN (?, ?) "
                 + "AND friend_id NOT IN (?, ?))";
         SqlRowSet srs = jdbcTemplate.queryForRowSet(sqlQuery, friend1, friend2, friend1, friend2);
@@ -120,26 +141,9 @@ public class UserDbStorage implements UserStorage {
 
     public boolean isFriend(int userId, int friendId) {
         String sqlQuery = "SELECT * FROM friends WHERE "
-                + "user_id = ? AND friends_id = ?";
+                + "user_id = ? AND friend_id = ?";
         SqlRowSet srs = jdbcTemplate.queryForRowSet(sqlQuery, userId, friendId);
         return srs.next();
-    }
-
-    private static User userMap(SqlRowSet srs) {
-        int id = srs.getInt("user_id");
-        String name = srs.getString("user_name");
-        String login = srs.getString("login");
-        String email = srs.getString("email");
-        LocalDate birthday = Objects.requireNonNull(srs.getTimestamp("birthday"))
-                .toLocalDateTime().toLocalDate();
-        User user = User.builder()
-                .id(id)
-                .name(name)
-                .login(login)
-                .email(email)
-                .birthday(birthday)
-                .build();
-        return user;
     }
 
     public void clearDbUsers() {
